% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVarEEclosestImage.R
\name{addVarEEclosestImage}
\alias{addVarEEclosestImage}
\title{Add variable from closest image in a Google Earth Engine Collection}
\usage{
addVarEEclosestImage(
  ee_counts,
  collection,
  reducer,
  maxdiff,
  bands = NULL,
  unmask = FALSE
)
}
\arguments{
\item{ee_counts}{A feature collection with the counts we want to annotate.
We need to upload an sf object with the spatial location of the counts to GEE.
This object must also have a character column with the dates that need to be
matched against the image collection dates. The format must be "yyyy-mm-dd"
and the column must be named "Date".}

\item{collection}{Either a character string with the name of the collection
we want to use or a GEE collection produced with \code{ee$ImageCollection()}.
See \href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{reducer}{A character string specifying the function to apply when
the location of the counts is a reference polygon and we need to extract a
summary of the images in 'collection'. It is common to use "mean", "sum" or
"count". But there are many other, see 'ee.Reducer' under Client Libraries at
\url{https://developers.google.com/earth-engine/apidocs}.}

\item{maxdiff}{Maximum difference in days allowed for an image to be matched with
data.}

\item{bands}{Select specific bands from the image. Only one band at a time is
allowed for now.}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}
}
\description{
Each feature in a feature collection (i.e. row in a simple feature
collection) is matched with the image in an image collection that is closest
to it in time.
}
\examples{
\dontrun{
# Load the remote data asset
ee_data <- ee$FeatureCollection(assetId)  # assetId must correspond to an asset in your GEE account

# Annotate with TerraClimate dataset
visit_new <- addVarEEclosestImage(ee_counts = ee_data,
                                  collection = "IDAHO_EPSCOR/TERRACLIMATE",
                                  reducer = "mean",                          # We only need spatial reducer
                                  maxdiff = 15,                              # This is the maximum time difference that GEE checks
                                  bands = c("tmmn"))
}
}
