% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwac_count_vars.R
\docType{data}
\name{cwac_count_vars}
\alias{cwac_count_vars}
\title{Empty data frame with CWAC counts variables}
\format{
A data frame with 0 rows and 39 variables:
\describe{
\item{item}{DESCRIPTION}
...
}
}
\usage{
cwac_count_vars
}
\description{
An empty dataset containing all variables that would be downloaded
for any species with data. This is useful for species that don't have any data
on CWAC but are still in the project species list.
}
\keyword{datasets}
