% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCwacSiteBoundary.R
\name{getCwacSiteBoundary}
\alias{getCwacSiteBoundary}
\title{Get a CWAC site boundary}
\usage{
getCwacSiteBoundary(loc_code)
}
\arguments{
\item{loc_code}{The code of a CWAC site}
}
\value{
A simple feature polygon for a CWAC site.
}
\description{
Get a CWAC site boundary
}
\details{
This function retrieves the coordinate list for a CWAC site and
creates an simple feature polygon.
}
\examples{
getCwacSiteBoundary(26352535)
getCwacSiteBoundary("26352535")
}
