% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCwacSiteInfo.R
\name{getCwacSiteInfo}
\alias{getCwacSiteInfo}
\title{Retrives available metadata for a CWAC site}
\usage{
getCwacSiteInfo(loc_code)
}
\arguments{
\item{loc_code}{Numeric or character string corresponding to a CWAC site code}
}
\value{
A tibble with metadata for a specific CWAC site
}
\description{
Retrives available metadata for a CWAC site
}
\examples{
getCwacSiteInfo(23312919)
getCwacSiteInfo("23312919")
}
