% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCwacSppList.R
\name{getCwacSppList}
\alias{getCwacSppList}
\title{Get CWAC species list}
\usage{
getCwacSppList()
}
\value{
A dataframe with a list of all species recorded in the CWAC project
}
\description{
Get CWAC species list
}
\examples{
getCwacSppList()
}
