% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCwacSurvey.R
\name{getCwacSurvey}
\alias{getCwacSurvey}
\title{Download information from a particular CWAC survey card}
\usage{
getCwacSurvey(card)
}
\arguments{
\item{card}{Numeric or character string correspoding to the card code to download information for}
}
\value{
A list with three elements: i) summary of the survey metadata, ii) information about the observers that collected the data
(if available), iii) counts recorded in the selected card
}
\description{
Download information from a particular CWAC survey card
}
\examples{
getCwacSurvey(508082)
getCwacSurvey("508082")
}
