% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCwacCards.R
\name{listCwacCards}
\alias{listCwacCards}
\title{List all cards submitted for a CWAC site}
\usage{
listCwacCards(loc_code)
}
\arguments{
\item{loc_code}{character string corresponding to the location code of the site of interest}
}
\value{
a tibble with a list of all cards submitted for a specific CWAC site
}
\description{
List all cards submitted for a CWAC site
}
\examples{
listCwacCards(32481810)
listCwacCards("32481810")
}
