% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCwacSites.R
\name{listCwacSites}
\alias{listCwacSites}
\title{Retrieve a list of active CWAC sites registered in a South African province}
\usage{
listCwacSites(.region_type = c("country", "province", "pentad"), .region)
}
\arguments{
\item{.region_type}{The type of region we are interested on.
Three options: "country", "province" and "pentad".}

\item{.region}{A character string corresponding to the specific region we are
interested in. It can be either a country in Southern Africa, a South African
province or a pentad code.}
}
\value{
A tibble with a list of all active sites for the province
}
\description{
Retrieve a list of active CWAC sites registered in a South African province
}
\examples{
listCwacSites(.region_type = "province", .region = "Western Cape")
listCwacSites(.region_type = "country", .region = "Kenya")
}
