% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCwacSpp.R
\name{listCwacSpp}
\alias{listCwacSpp}
\title{Get CWAC species list}
\usage{
listCwacSpp()
}
\value{
A dataframe with a list of all species recorded in the CWAC project
}
\description{
Get CWAC species list
}
\examples{
listCwacSpp()
}
