% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCwacSppCounts.R
\name{getCwacSppCounts}
\alias{getCwacSppCounts}
\title{Get all CWAC counts for a species}
\usage{
getCwacSppCounts(spp_code)
}
\arguments{
\item{spp_code}{The CWAC code of a species. Only a single species is accepted
at the moment.}
}
\value{
A tibble with all the counts recorded for the desired species.
Note that there might be warnings about formatting.
}
\description{
Get all CWAC counts for a species
}
\examples{
getCwacSppCounts(41)
getCwacSppCounts("41")
}
