% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCwacSites.R
\name{listCwacSites}
\alias{listCwacSites}
\title{Retrieve a list of active CWAC sites registered in a South African province}
\usage{
listCwacSites(province)
}
\arguments{
\item{province}{A character string with the name of the province of interest.
Available provinces are:
Northern Province, Mpumalanga, North West, Gauteng, KwaZulu-Natal,
Free State, Northern Cape, Western Cape, Eastern Cape, Kenya, Angola,
Tanzania, Limpopo.
NORTHERN PROVINCE IS ACTUALLY NOT AVAILABLE, ALTHOUGH IT SAYS IT IS ON THE
WEBSITE.}
}
\value{
A tibble with a list of all active sites for the province
}
\description{
Retrieve a list of active CWAC sites registered in a South African province
}
\examples{
listCwacSites("Western Cape")
listCwacSites("western cape")
}
