% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCwacTerm.R
\name{searchCwacTerm}
\alias{searchCwacTerm}
\title{Search for a term in the CWAC dicctionary}
\usage{
searchCwacTerm(term = NULL, option = c("term", "fields"))
}
\arguments{
\item{term}{A character string corresponding a term found in CWAC data. The term must be spelled exactly as it appears in the data.}

\item{option}{A character string with one of two options: "term" will search for the term specified in the term argument, "fields" returns a vector with the names of all possible search terms}
}
\value{
A list with a description of the term and, if it exists, the different values that the variable can take on.
}
\description{
Search for a term in the CWAC dicctionary
}
\examples{
searchCwacTerm(term = "Season")
searchCwacTerm(option = "fields")
}
